PUSHING SOURCE CODE TO GIT

+++ Jenkinsfile
@@ -25,13 +25,17 @@ pipeline {
 
                     sh '''
                     ContainerId=$(docker ps -q)
-                    if [ -n '$ContainerId' ]; then
+                    ImageId=$(docker images -q)
+                    if [ -n '$ContainerId' ] && [ -n "$ImageId" ]; then
                         echo "STOPPING THE CONTAINER: $ContainerId"
                         docker stop $ContainerId
+                        echo "IMAGE ID IS: $ImageId
+                        docker rmi $ImageId
 
                         if [ $? -eq 0 ]; then
                             echo "REMOVING THE CONATINER: $ContainerId"
                             docker rm $ContainerId
+                            
                         fi
                     else
                         echo "NO RUNNING CONTAINER!!!"
